refactor: move schema code to kernel module
