+++ crates/core/src/kernel/schema/partitions.rs
@@ -6,8 +6,9 @@ use std::convert::TryFrom;
 use delta_kernel::expressions::Scalar;
 use serde::{Serialize, Serializer};
 
+use super::{DataType, PrimitiveType};
 use crate::errors::DeltaTableError;
-use crate::kernel::{scalars::ScalarExt, DataType, PrimitiveType};
+use crate::kernel::scalars::ScalarExt;
 
 /// A special value used in Hive to represent the null partition in partitioned tables
 pub const NULL_PARTITION_VALUE_DATA_PATH: &str = "__HIVE_DEFAULT_PARTITION__";
