+++ crates/core/src/operations/mod.rs
@@ -40,7 +40,6 @@ use crate::DeltaTable;
 
 pub mod add_column;
 pub mod add_feature;
-pub mod cast;
 pub mod convert_to_delta;
 pub mod create;
 pub mod drop_constraints;
