--- crates/core/src/schema/mod.rs
@@ -1,2 +1,0 @@
-//! Delta Table schema implementation.
-pub mod partitions;
