+++ crates/core/src/writer/record_batch.rs
@@ -29,9 +29,9 @@ use super::utils::{
 };
 use super::{DeltaWriter, DeltaWriterError, WriteMode};
 use crate::errors::DeltaTableError;
+use crate::kernel::schema::merge_arrow_schema;
 use crate::kernel::{scalars::ScalarExt, Action, Add, PartitionsExt};
 use crate::logstore::ObjectStoreRetryExt;
-use crate::operations::cast::merge_schema::merge_arrow_schema;
 use crate::table::builder::DeltaTableBuilder;
 use crate::table::config::DEFAULT_NUM_INDEX_COLS;
 use crate::DeltaTable;
